+++ releases.md
@@ -3,6 +3,12 @@
 jexler releases
 ===============
 
+next
+----
+
+* CronService: Cron string "now+stop" for a single event immediately
+  followed by a single StopEvent.
+
 1.0.4 (2013-07-23)
 ------------------